change: rename sqlc_db module to database
